# sink(ticdc): optimize buffer sink flush from O(N^2) to O(N)