--- a/cdc/sink/buffer_sink.go
+++ b/cdc/sink/buffer_sink.go
@@ -24,8 +24,11 @@ import (
 	"github.com/pingcap/log"
 	"github.com/pingcap/ticdc/cdc/model"
 	"github.com/pingcap/ticdc/pkg/util"
+	"github.com/prometheus/client_golang/prometheus"
 	"go.uber.org/zap"
 )
+
+const maxFlushBatchSize = 512
 
 type bufferSink struct {
 	Sink
@@ -38,99 +41,127 @@ type bufferSink struct {
 }
 
 func newBufferSink(
-	ctx context.Context,
-	backendSink Sink,
-	errCh chan error,
-	checkpointTs model.Ts,
-	drawbackChan chan drawbackMsg,
+	backendSink Sink, checkpointTs model.Ts, drawbackChan chan drawbackMsg,
 ) *bufferSink {
 	sink := &bufferSink{
 		Sink: backendSink,
 		// buffer shares the same flow control with table sink
 		buffer:                 make(map[model.TableID][]*model.RowChangedEvent),
 		changeFeedCheckpointTs: checkpointTs,
-		flushTsChan:            make(chan flushMsg, 128),
+		flushTsChan:            make(chan flushMsg, maxFlushBatchSize),
 		drawbackChan:           drawbackChan,
 	}
-	go sink.run(ctx, errCh)
 	return sink
+}
+
+type runState struct {
+	batch [maxFlushBatchSize]flushMsg
+
+	metricFlushDuration   prometheus.Observer
+	metricEmitRowDuration prometheus.Observer
+	metricBufferSize      prometheus.Counter
+	metricTotalRows       prometheus.Counter
 }
 
 func (b *bufferSink) run(ctx context.Context, errCh chan error) {
 	changefeedID := util.ChangefeedIDFromCtx(ctx)
 	advertiseAddr := util.CaptureAddrFromCtx(ctx)
-	metricFlushDuration := flushRowChangedDuration.WithLabelValues(advertiseAddr, changefeedID, "Flush")
-	metricEmitRowDuration := flushRowChangedDuration.WithLabelValues(advertiseAddr, changefeedID, "EmitRow")
-	metricBufferSize := bufferChanSizeGauge.WithLabelValues(advertiseAddr, changefeedID)
-	metricTotalRows := bufferSinkTotalRowsCountCounter.WithLabelValues(advertiseAddr, changefeedID)
+	state := runState{
+		metricFlushDuration:   flushRowChangedDuration.WithLabelValues(advertiseAddr, changefeedID, "Flush"),
+		metricEmitRowDuration: flushRowChangedDuration.WithLabelValues(advertiseAddr, changefeedID, "EmitRow"),
+		metricTotalRows:       bufferSinkTotalRowsCountCounter.WithLabelValues(advertiseAddr, changefeedID),
+	}
 	defer func() {
 		flushRowChangedDuration.DeleteLabelValues(advertiseAddr, changefeedID, "Flush")
 		flushRowChangedDuration.DeleteLabelValues(advertiseAddr, changefeedID, "EmitRow")
-		bufferChanSizeGauge.DeleteLabelValues(advertiseAddr, changefeedID)
 		bufferSinkTotalRowsCountCounter.DeleteLabelValues(advertiseAddr, changefeedID)
 	}()
+
 	for {
-		select {
-		case <-ctx.Done():
-			err := ctx.Err()
-			if err != nil && errors.Cause(err) != context.Canceled {
-				errCh <- err
-			}
+		keepRun, err := b.runOnce(ctx, &state)
+		if err != nil && errors.Cause(err) != context.Canceled {
+			errCh <- err
 			return
-		case drawback := <-b.drawbackChan:
-			b.bufferMu.Lock()
-			delete(b.buffer, drawback.tableID)
-			b.bufferMu.Unlock()
-			close(drawback.callback)
-		case flushEvent := <-b.flushTsChan:
-			b.bufferMu.Lock()
-			resolvedTs := flushEvent.resolvedTs
-			// find all rows before resolvedTs and emit to backend sink
-			for tableID, rows := range b.buffer {
-				i := sort.Search(len(rows), func(i int) bool {
-					return rows[i].CommitTs > resolvedTs
-				})
-				metricTotalRows.Add(float64(i))
-
-				start := time.Now()
-				err := b.Sink.EmitRowChangedEvents(ctx, rows[:i]...)
-				if err != nil {
-					b.bufferMu.Unlock()
-					if errors.Cause(err) != context.Canceled {
-						errCh <- err
-					}
-					return
-				}
-				dur := time.Since(start)
-				metricEmitRowDuration.Observe(dur.Seconds())
-
-				// put remaining rows back to buffer
-				// append to a new, fixed slice to avoid lazy GC
-				b.buffer[tableID] = append(make([]*model.RowChangedEvent, 0, len(rows[i:])), rows[i:]...)
-			}
-			b.bufferMu.Unlock()
-
-			start := time.Now()
-			tableID := flushEvent.tableID
-			checkpointTs, err := b.Sink.FlushRowChangedEvents(ctx, flushEvent.tableID, resolvedTs)
-			if err != nil {
-				if errors.Cause(err) != context.Canceled {
-					errCh <- err
-				}
-				return
-			}
-			b.tableCheckpointTsMap.Store(tableID, checkpointTs)
-
-			dur := time.Since(start)
-			metricFlushDuration.Observe(dur.Seconds())
-			if dur > 3*time.Second {
-				log.Warn("flush row changed events too slow",
-					zap.Duration("duration", dur), util.ZapFieldChangefeed(ctx))
-			}
-		case <-time.After(defaultMetricInterval):
-			metricBufferSize.Set(float64(len(b.buffer)))
+		}
+		if !keepRun {
+			return
 		}
 	}
+}
+
+func (b *bufferSink) runOnce(ctx context.Context, state *runState) (bool, error) {
+	batchSize, batch := 0, state.batch
+	push := func(event flushMsg) {
+		batch[batchSize] = event
+		batchSize++
+	}
+	select {
+	case <-ctx.Done():
+		return false, ctx.Err()
+	case drawback := <-b.drawbackChan:
+		b.bufferMu.Lock()
+		delete(b.buffer, drawback.tableID)
+		b.bufferMu.Unlock()
+		close(drawback.callback)
+	case event := <-b.flushTsChan:
+		push(event)
+	RecvBatch:
+		for batchSize < maxFlushBatchSize {
+			select {
+			case event := <-b.flushTsChan:
+				push(event)
+			default:
+				break RecvBatch
+			}
+		}
+	}
+
+	b.bufferMu.Lock()
+	startEmit := time.Now()
+	// find all rows before resolvedTs and emit to backend sink
+	for i := 0; i < batchSize; i++ {
+		tableID, resolvedTs := batch[i].tableID, batch[i].resolvedTs
+		rows := b.buffer[tableID]
+		i := sort.Search(len(rows), func(i int) bool {
+			return rows[i].CommitTs > resolvedTs
+		})
+		if i == 0 {
+			continue
+		}
+		state.metricTotalRows.Add(float64(i))
+
+		err := b.Sink.EmitRowChangedEvents(ctx, rows[:i]...)
+		if err != nil {
+			b.bufferMu.Unlock()
+			return false, errors.Trace(err)
+		}
+
+		// put remaining rows back to buffer
+		// append to a new, fixed slice to avoid lazy GC
+		b.buffer[tableID] = append(make([]*model.RowChangedEvent, 0, len(rows[i:])), rows[i:]...)
+	}
+	b.bufferMu.Unlock()
+	state.metricEmitRowDuration.Observe(time.Since(startEmit).Seconds())
+
+	startFlush := time.Now()
+	for i := 0; i < batchSize; i++ {
+		tableID, resolvedTs := batch[i].tableID, batch[i].resolvedTs
+		checkpointTs, err := b.Sink.FlushRowChangedEvents(ctx, tableID, resolvedTs)
+		if err != nil {
+			return false, errors.Trace(err)
+		}
+		b.tableCheckpointTsMap.Store(tableID, checkpointTs)
+	}
+	now := time.Now()
+	state.metricFlushDuration.Observe(now.Sub(startFlush).Seconds())
+	if now.Sub(startEmit) > time.Second {
+		log.Warn("flush row changed events too slow",
+			zap.Int("batchSize", batchSize),
+			zap.Duration("duration", now.Sub(startEmit)),
+			util.ZapFieldChangefeed(ctx))
+	}
+
+	return true, nil
 }
 
 func (b *bufferSink) EmitRowChangedEvents(ctx context.Context, rows ...*model.RowChangedEvent) error {