--- a/cdc/sink/manager.go
+++ b/cdc/sink/manager.go
@@ -40,9 +40,6 @@ type Manager struct {
 	tableSinksMu           sync.Mutex
 	changeFeedCheckpointTs uint64
 
-	flushMu  sync.Mutex
-	flushing int64
-
 	drawbackChan chan drawbackMsg
 
 	captureAddr               string
@@ -56,8 +53,10 @@ func NewManager(
 	captureAddr string, changefeedID model.ChangeFeedID,
 ) *Manager {
 	drawbackChan := make(chan drawbackMsg, 16)
+	bufSink := newBufferSink(backendSink, checkpointTs, drawbackChan)
+	go bufSink.run(ctx, errCh)
 	return &Manager{
-		backendSink:               newBufferSink(ctx, backendSink, errCh, checkpointTs, drawbackChan),
+		backendSink:               bufSink,
 		changeFeedCheckpointTs:    checkpointTs,
 		tableSinks:                make(map[model.TableID]*tableSink),
 		drawbackChan:              drawbackChan,
@@ -110,20 +109,8 @@ func (m *Manager) getMinEmittedTs(tableID model.TableID) model.Ts {
 	return minTs
 }
 
-func (m *Manager) flushBackendSink(ctx context.Context, tableID model.TableID) (model.Ts, error) {
-	// NOTICE: Because all table sinks will try to flush backend sink,
-	// which will cause a lot of lock contention and blocking in high concurrency cases.
-	// So here we use flushing as a lightweight lock to improve the lock competition problem.
-	if !atomic.CompareAndSwapInt64(&m.flushing, 0, 1) {
-		return m.getCheckpointTs(tableID), nil
-	}
-	m.flushMu.Lock()
-	defer func() {
-		m.flushMu.Unlock()
-		atomic.StoreInt64(&m.flushing, 0)
-	}()
-	minEmittedTs := m.getMinEmittedTs(tableID)
-	checkpointTs, err := m.backendSink.FlushRowChangedEvents(ctx, tableID, minEmittedTs)
+func (m *Manager) flushBackendSink(ctx context.Context, tableID model.TableID, resolvedTs uint64) (model.Ts, error) {
+	checkpointTs, err := m.backendSink.FlushRowChangedEvents(ctx, tableID, resolvedTs)
 	if err != nil {
 		return m.getCheckpointTs(tableID), errors.Trace(err)
 	}