--- a/cdc/sink/table_sink.go
+++ b/cdc/sink/table_sink.go
@@ -15,20 +15,19 @@ package sink
 
 import (
 	"context"
+	"log"
 	"sort"
-	"sync/atomic"
 
 	"github.com/pingcap/errors"
 	"github.com/pingcap/ticdc/cdc/model"
 	"github.com/pingcap/ticdc/cdc/redo"
+	"go.uber.org/zap"
 )
 
 type tableSink struct {
-	tableID model.TableID
-	manager *Manager
-	buffer  []*model.RowChangedEvent
-	// emittedTs means all of events which of commitTs less than or equal to emittedTs is sent to backendSink
-	emittedTs   model.Ts
+	tableID     model.TableID
+	manager     *Manager
+	buffer      []*model.RowChangedEvent
 	redoManager redo.LogManager
 }
 
@@ -50,16 +49,15 @@ func (t *tableSink) EmitDDLEvent(ctx context.Context, ddl *model.DDLEvent) error
 // is required to be no more than global resolvedTs, table barrierTs and table
 // redo log watermarkTs.
 func (t *tableSink) FlushRowChangedEvents(ctx context.Context, tableID model.TableID, resolvedTs uint64) (uint64, error) {
+	if tableID != t.tableID {
+		log.Panic("inconsistent table sink",
+			zap.Int64("tableID", tableID), zap.Int64("sinkTableID", t.tableID))
+	}
 	i := sort.Search(len(t.buffer), func(i int) bool {
 		return t.buffer[i].CommitTs > resolvedTs
 	})
 	if i == 0 {
-		atomic.StoreUint64(&t.emittedTs, resolvedTs)
-		ckpt, err := t.flushRedoLogs(ctx, resolvedTs)
-		if err != nil {
-			return ckpt, err
-		}
-		return t.manager.flushBackendSink(ctx, tableID, resolvedTs)
+		return t.flushResolvedTs(ctx, resolvedTs)
 	}
 	resolvedRows := t.buffer[:i]
 	t.buffer = append(make([]*model.RowChangedEvent, 0, len(t.buffer[i:])), t.buffer[i:]...)
@@ -68,36 +66,31 @@ func (t *tableSink) FlushRowChangedEvents(ctx context.Context, tableID model.Tab
 	if err != nil {
 		return t.manager.getCheckpointTs(tableID), errors.Trace(err)
 	}
-	atomic.StoreUint64(&t.emittedTs, resolvedTs)
-	ckpt, err := t.flushRedoLogs(ctx, resolvedTs)
-	if err != nil {
-		return ckpt, err
-	}
-	return t.manager.flushBackendSink(ctx, tableID, resolvedTs)
+	return t.flushResolvedTs(ctx, resolvedTs)
 }
 
+func (t *tableSink) flushResolvedTs(ctx context.Context, resolvedTs uint64) (uint64, error) {
+	redoTs, err := t.flushRedoLogs(ctx, resolvedTs)
+	if err != nil {
+		return t.manager.getCheckpointTs(t.tableID), err
+	}
+	if redoTs < resolvedTs {
+		resolvedTs = redoTs
+	}
+	return t.manager.flushBackendSink(ctx, t.tableID, resolvedTs)
+}
+
+// flushRedoLogs flush redo logs and returns redo log resolved ts which means
+// all events before the ts have been persisted to redo log storage.
 func (t *tableSink) flushRedoLogs(ctx context.Context, resolvedTs uint64) (uint64, error) {
 	if t.redoManager.Enabled() {
 		err := t.redoManager.FlushLog(ctx, t.tableID, resolvedTs)
 		if err != nil {
-			return t.manager.getCheckpointTs(t.tableID), err
+			return 0, err
 		}
+		return t.redoManager.GetMinResolvedTs(), nil
 	}
-	return 0, nil
-}
-
-// getResolvedTs returns resolved ts, which means all events before resolved ts
-// have been sent to sink manager, if redo log is enabled, it also means all events
-// before resolved ts have been persisted to redo log storage.
-func (t *tableSink) getResolvedTs() uint64 {
-	ts := atomic.LoadUint64(&t.emittedTs)
-	if t.redoManager.Enabled() {
-		redoResolvedTs := t.redoManager.GetMinResolvedTs()
-		if redoResolvedTs < ts {
-			ts = redoResolvedTs
-		}
-	}
-	return ts
+	return resolvedTs, nil
 }
 
 func (t *tableSink) EmitCheckpointTs(ctx context.Context, ts uint64) error {