@overvenus 2021-12-16
> Can you add some description about how it works?

Updated, see description.